+++ offload/src/Kernels/Sanitizer.cpp
@@ -11,11 +11,11 @@
 #include <cstdint>
 
 extern "C" {
-__device__ void __offload_san_register_host(void *Ptr, uint32_t Size,
+__device__ void __offload_san_register_host(void *Ptr, uint64_t Size,
                                             uint32_t SlotId);
 __device__ void __offload_san_unregister_host(void *Ptr);
 __device__ void __offload_san_get_ptr_info(uint32_t SlotId, void **Ptr,
-                                           uint32_t *Size,
+                                           uint64_t *Size,
                                            uint64_t *LocationIdPtr);
 __device__ void __offload_san_get_global_info(void *Ptr,
                                               uint64_t *LocationIdPtr);
@@ -31,7 +31,7 @@ __sanitizer_unregister(void *P) {
 }
 
 [[gnu::weak, clang::disable_sanitizer_instrumentation]] __global__ void
-__sanitizer_get_ptr_info(uint32_t SlotId, void **Ptr, uint32_t *Size,
+__sanitizer_get_ptr_info(uint32_t SlotId, void **Ptr, uint64_t *Size,
                          uint64_t *LocationIdPtr) {
   __offload_san_get_ptr_info(SlotId, Ptr, Size, LocationIdPtr);
 }
